+++ subprojects/file-watching/src/main/java/org/gradle/internal/watch/vfs/impl/WatchingNotSupportedFileSystemWatchingHandler.java
@@ -16,8 +16,8 @@
 
 package org.gradle.internal.watch.vfs.impl;
 
-import org.gradle.internal.snapshot.VfsRoot;
-import org.gradle.internal.snapshot.VfsRootReference;
+import org.gradle.internal.vfs.VfsRoot;
+import org.gradle.internal.vfs.VfsRootReference;
 import org.gradle.internal.watch.vfs.FileSystemWatchingHandler;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
